refactor eval harness: extract LiteLLM/ToolProvider wiring, make scratch always-on
- Extract tool_params_from_provider and tool_result_content into
  litellm_tool_provider.py (separate from harness.py)
- Remove dead Recommendation model and recommendations field from RunSummary
- Make agent_tool_provider required (not optional) in _TwentyQuestionsRunner
  and run_twenty_questions
- Always enable scratch container (remove --scratch flag)
- Remove testonly from twenty_questions library/binary and docker_scratch;
  remove testonly from util:oci (no testonly deps, useful outside tests)
- Make conftest a separate py_library target
- Fix end_game tool to return None instead of {"status": "game_ended"}
