+++ nix/home/skills/info_gathering/evals/twenty_questions/BUILD.bazel
@@ -2,13 +2,13 @@ load("@rules_python//python:defs.bzl", "py_binary", "py_library", "py_test")
 
 py_library(
     name = "twenty_questions",
-    testonly = True,
     srcs = ["twenty_questions.py"],
     imports = ["../../../../../.."],
     deps = [
         "//agent_core:tool_provider",
         "//nix/home/skills/info_gathering/evals:docker_scratch",
         "//nix/home/skills/info_gathering/evals:harness",
+        "//nix/home/skills/info_gathering/evals:litellm_tool_provider",
         "//util/bazel:runfiles",
         "@pypi//litellm",
         "@pypi//pydantic",
@@ -17,7 +17,6 @@ py_library(
 
 py_binary(
     name = "twenty_questions_bin",
-    testonly = True,
     data = [
         "sim.txt",
         "//nix/home/skills/info_gathering:SKILL.md",
@@ -28,15 +27,21 @@ py_binary(
     deps = [":twenty_questions"],
 )
 
+py_library(
+    name = "conftest",
+    srcs = ["conftest.py"],
+    imports = ["../../../../../.."],
+    deps = ["@pypi//pytest"],
+)
+
 py_test(
     name = "test_twenty_questions",
-    srcs = [
-        "conftest.py",
-        "test_twenty_questions.py",
-    ],
+    srcs = ["test_twenty_questions.py"],
     imports = ["../../../../../.."],
     deps = [
+        ":conftest",
         ":twenty_questions",
+        "//agent_core:direct_provider",
         "//nix/home/skills/info_gathering/evals:harness",
         "@pypi//litellm",
         "@pypi//pytest_asyncio",
